requirements.txt
Pillow>=7.2.0
PyPDF2>=1.26.0
reportlab>=3.5.49
tk>=0.1.0
